moved ColorMapPickerCard.svelte
