+++ src/components/controls/ColorMapPickerCard.svelte
